+++ src/main/java/racingcar/Application.java
@@ -1,5 +1,7 @@
 package racingcar;
 
+import racingcar.Controller.Controller;
+
 public class Application {
     public static void main(String[] args) {
         Controller controller = new Controller();
