+++ src/main/java/racingcar/Domain/ProgressBoard.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.Domain;
 
 import java.util.HashMap;
 import java.util.List;
